Trait pushgen::Generator

pub trait Generator {
    type Output;
    fn run(
        &mut self, 
        output: impl FnMut(Self::Output) -> ValueResult
    ) -> GeneratorResult;
}

Trait for generating values into a closure.

When a Generator is run() it generates values that are fed an output closure. It continues to feed values to the closure for as long as it can, unless the closure returns ValueResult::Stop.

When all values have been generated the run() method returns GeneratorResult::Complete. If output returns ValueResult::Stop for any value the generator must not call output with any further values and return GeneratorResult::Stopped as well.

The generator must not assume that it won’t be called again after it returns.

Example

A generic generator can be written like this:

use pushgen::{Generator, ValueResult, GeneratorResult};
struct GenericGenerator<Out, Gen>
where
    Gen: FnMut() -> Option<Out>,
{
    generator: Gen,
}

impl<Out, Gen> Generator for GenericGenerator<Out, Gen>
    where
        Gen: FnMut() -> Option<Out>,
{
    type Output = Out;

    fn run(&mut self, mut output: impl FnMut(Self::Output) -> ValueResult) -> GeneratorResult {
        while let Some(value) = (self.generator)() {
            if output(value) == ValueResult::Stop {
                return GeneratorResult::Stopped;
            }
        }
        GeneratorResult::Complete
    }
}

Associated Types

type Output

Data-type generated by the generator.

Loading content...
Required methods

fn run(
&mut self,
output: impl FnMut(Self::Output) -> ValueResult
) -> GeneratorResult

Run the generator, emitting values to the output closure. New values are emitted for as long as the closure returns ValueResult::MoreValues. If the closure returns ValueResult::Stop the generator must return GeneratorResult::Stopped.

Loading content...