Trait pushgen::IntoGenerator

pub trait IntoGenerator {
    type Output;
    type IntoGen: Generator<Output = Self::Output>;
    fn into_gen(self) -> Self::IntoGen;
}

Conversion to a generator.

By implementing IntoGenerator for a type you define how that type will be converted to a generator.

Associated Types

type Output

Data-type generated by the generator.

type IntoGen: Generator<Output = Self::Output>

Which kind of generator are we turning this into?

Required methods

fn into_gen(self) -> Self::IntoGen

Creates a generator from a value.

See the module-level documentation for more.

Examples

Basic usage:

use pushgen::IntoGenerator;
use crate::pushgen::GeneratorExt;
let v = vec![1, 2, 3];
let mut gen = v.into_gen();

let mut output: Vec<i32> = Vec::new();
gen.for_each(|x| output.push(x));
assert_eq!(output, [1, 2, 3]);