Module pushgen::traits

Module containing the various traits used by pushgen.

Traits

DynGenerator

Sealed trait to enable boxed generators. See .boxed() for details.

FromGenerator

Conversion from a Generator.

Generator

Trait for generating values into a closure.

GeneratorExt

Provides extension-methods for all generators.

IntoGenerator

Conversion to a generator.

Product

Trait to represent types that can be created by multiplying values from a generator.

ReverseGenerator

A generator able to produce values from in reverse order.

Sum

Trait to represent types that can be created by summing up a generator.