Trait pushgen::traits::Product

pub trait Product<A = Self>: Sized {
    // Required method
    fn product<G>(gen: G) -> Self
       where G: Generator<Output = A>;
}

Trait to represent types that can be created by multiplying values from a generator.

The trait is used to implement the product() method on generators. Types which implement this trait can be generated by the product() method. This trait is generally interacted with via GeneratorExt::product.

Required Methods

fn product<G>(gen: G) -> Selfwhere G: Generator<Output = A>,

Calculate the product using the given generator.