Trait pushgen::traits::Sum

pub trait Sum<A = Self>: Sized {
    // Required method
    fn sum<G>(gen: G) -> Self
       where G: Generator<Output = A>;
}

Trait to represent types that can be created by summing up a generator.

The trait is used to implement the sum() method on generators. Types which implement this trait can be generated by the sum() method. This trait is generally interacted with via GeneratorExt::sum.

Required Methods

fn sum<G>(gen: G) -> Selfwhere G: Generator<Output = A>,

Calculate the sum from a given generator.