Trait pushgen::GeneratorExt

pub trait GeneratorExt: Sealed + Generator {

    fn last(self) -> Option<Self::Output>
    where
        Self: Sized,
    { ... }

    fn cloned<'a, T>(self) -> Cloned<Self>
    where
        Self: Generator<Output = &'a T> + Sized,
        T: 'a + Clone,
    { ... }

    fn chain<Gen>(self, other: Gen) -> Chain<Self, Gen>
    where
        Self: Sized,
        Gen: Generator<Output = Self::Output>,
    { ... }

    fn filter<Pred>(self, predicate: Pred) -> Filter<Self, Pred>
    where
        Self: Sized,
        Pred: FnMut(&Self::Output) -> bool,
    { ... }

    fn filter_map<B, F>(self, f: F) -> FilterMap<Self, F>
    where
        Self: Sized,
        F: FnMut(Self::Output) -> Option<B>,
    { ... }

    fn map<Trans, Out>(self, transform_fn: Trans) -> Map<Self, Trans>
    where
        Self: Sized,
        Trans: FnMut(Self::Output) -> Out,
    { ... }

    fn skip(self, n: usize) -> Skip<Self>
    where
        Self: Sized,
    { ... }

    fn take(self, n: usize) -> Take<Self>
    where
        Self: Sized,
    { ... }

    fn flat_map<U, F>(self, f: F) -> Flatten<Map<Self, F>>
    where
        Self: Sized,
        U: IntoGenerator,
        F: FnMut(Self::Output) -> U,
    { ... }

    fn flatten(self) -> Flatten<Self>
    where
        Self: Sized,
        Self::Output: IntoGenerator,
    { ... }

    fn for_each<Func>(&mut self, func: Func) -> GeneratorResult
    where
        Self: Sized,
        Func: FnMut(Self::Output),
    { ... }

    fn try_for_each<F, E>(&mut self, f: F) -> Result<(), E>
    where
        Self: Sized,
        F: FnMut(Self::Output) -> Result<(), E>,
    { ... }

    fn zip<Right>(self, right: Right) -> Zip<Self, Right>
    where
        Self: Sized,
        Right: Generator,
    { ... }

    fn dedup(self) -> Dedup<Self>
    where
        Self: Sized,
        Self::Output: PartialEq,
    { ... }

    fn iter(self) -> IteratorAdaptor<Self>
    where
        Self: Sized,
    { ... }

    fn boxed(self) -> BoxedGenerator<Self::Output>
    where
        Self: Sized + 'static,
    { ... }

}

Provides extension-methods for all generators.

This allows generators to be composed to new generators, or consumed.

Example

use pushgen::{SliceGenerator, GeneratorExt};
let data = [1, 2, 3, 4];
let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&data).map(|x| x*3).for_each(|x| output.push(x));
assert_eq!(output, [3,6,9,12]);

Provided methods

fn last(self) -> Option<Self::Output> where
Self: Sized,

Exhausts the generator, returning the last element.

This method will evaluate the generator until it completes. While doing so, it keeps track of the current element. After it completes last() will then return the last element it saw.

Examples

Basic usage:

use pushgen::IntoGenerator;
use pushgen::GeneratorExt;

let a = [1, 2, 3];
assert_eq!(a.into_gen().last(), Some(&3));

let a = [1, 2, 3, 4, 5];
assert_eq!(a.into_gen().last(), Some(&5));

fn cloned<'a, T>(self) -> Cloned<Self> where
Self: Generator<Output = &'a T> + Sized,
T: 'a + Clone,

Creates a generator that clones all of its elements.

This is useful when you have a generator that generates &T but you need a generate that generates T.

Examples

Basic usage

use pushgen::{SliceGenerator, GeneratorExt};
let data = [1, 2, 3];
let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&data).cloned().for_each(|x| output.push(x));
assert_eq!(output, [1, 2, 3])

fn chain<Gen>(self, other: Gen) -> Chain<Self, Gen> where
Self: Sized,
Gen: Generator<Output = Self::Output>,

Creates a generator by chaining two generators, running them one after the other.

Example

use pushgen::{SliceGenerator, GeneratorExt};
let data = [1, 2, 3];
let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&data).chain(SliceGenerator::new(&data)).for_each(|x| output.push(*x));
assert_eq!(output, [1, 2, 3, 1, 2, 3]);

fn filter<Pred>(self, predicate: Pred) -> Filter<Self, Pred> where
Self: Sized,
Pred: FnMut(&Self::Output) -> bool,

Create a filtered generator. Only values for which the predicate returns true will be passed on.

The predicate must implement FnMut(&Gen::Output) -> bool.

Example

let input = [1,2,3,4];
let mut output: Vec<i32> = Vec::new();
let run_result = SliceGenerator::new(&input).filter(|x| *x % 2 == 0).for_each(|x| output.push(*x));
assert_eq!(run_result, GeneratorResult::Complete);
assert_eq!(output, [2,4]);

fn filter_map<B, F>(self, f: F) -> FilterMap<Self, F> where
Self: Sized,
F: FnMut(Self::Output) -> Option<B>,

Creates a generator that both filters and maps.

The returned generator produces only the values for which the supplied closure returns Some(value).

filter_map can be used to make chains of filter and map more concise. The example below shows how a map().filter().map() can be shortened to a single call to filter_map.

Examples

Basic usage:

use pushgen::{SliceGenerator, GeneratorExt};

let a = ["1", "two", "NaN", "four", "5"];

let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&a).filter_map(|s| s.parse().ok()).for_each(|x: i32| output.push(x));
assert_eq!(output, [1, 5]);

Here’s the same example, but with filter and map:

use pushgen::{SliceGenerator, GeneratorExt};

let a = ["1", "two", "NaN", "four", "5"];

let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&a).map(|s| s.parse()).filter(|s| s.is_ok()).map(|s| s.unwrap()).for_each(|x: i32| output.push(x));
assert_eq!(output, [1, 5]);

fn map<Trans, Out>(self, transform_fn: Trans) -> Map<Self, Trans> where
Self: Sized,
Trans: FnMut(Self::Output) -> Out,

Takes a closure and creates a generator which calls the closure on each value.

Example

use pushgen::{SliceGenerator, GeneratorExt};
let data = [1, 2, 3];
let mut output: Vec<String> = Vec::new();
SliceGenerator::new(&data).map(|x| x.to_string()).for_each(|x| output.push(x));
assert_eq!(output, ["1", "2", "3"]);

fn skip(self, n: usize) -> Skip<Self> where
Self: Sized,

Skips over n values, consuming and ignoring them.

Example

 let input = [1,2,3,4];
 let mut skipped_generator = SliceGenerator::new(&input).skip(2);
 let mut output: Vec<i32> = Vec::new();
 skipped_generator.for_each(|x| output.push(*x));
 assert_eq!(output, [3,4]);

fn take(self, n: usize) -> Take<Self> where
Self: Sized,

Takes n values and then completes the generator.

Example

use pushgen::{SliceGenerator, GeneratorExt};
let data = [1, 2, 3, 4];
let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&data).take(2).for_each(|x| output.push(*x));
assert_eq!(output, [1, 2]);

fn flat_map<U, F>(self, f: F) -> Flatten<Map<Self, F>> where
Self: Sized,
U: IntoGenerator,
F: FnMut(Self::Output) -> U,

Creates a generator that works like map, but flattens nested structure.

The map adapter is very useful, but only when the closure argument produces values. If it produces a generator instead, there’s an extra layer of indirection. flat_map() will remove this extra layer on its own.

You can think of flat_map(f) as the semantic equivalent of mapping, and then flattening as in map(f).flatten().

Another way of thinking about flat_map(): map’s closure returns one item for each element, and flat_map()’s closure returns an iterator for each element.

Examples

Basic usage:

use pushgen::IntoGenerator;
use crate::pushgen::GeneratorExt;

let words = ["alpha", "beta", "gamma"];

let mut merged = String::new();
words.into_gen()
     .flat_map(|s| pushgen::from_iter(s.chars()))
     .for_each(|x| merged.push(x));
assert_eq!(merged, "alphabetagamma");

fn flatten(self) -> Flatten<Self> where
Self: Sized,
Self::Output: IntoGenerator,

Creates a generator that flattens nested structure.

This is useful when you have a generator of generators or a generator of things that can be turned into generators and you want to remove one level of indirection.

Examples

Basic usage:

use pushgen::IntoGenerator;
use crate::pushgen::GeneratorExt;

let data = vec![vec![1, 2, 3, 4], vec![5, 6]];
let mut output: Vec<i32> = Vec::new();
let flattened = data.into_gen().flatten().for_each(|x| output.push(x));
assert_eq!(output, [1, 2, 3, 4, 5, 6]);

Mapping and then flattening:

use pushgen::IntoGenerator;
use crate::pushgen::GeneratorExt;

let words = &["alpha", "beta", "gamma"];

let mut merged = String::new();
words.into_gen()
     .map(|s| pushgen::from_iter(s.chars()))
     .flatten()
     .for_each(|x| merged.push(x));
assert_eq!(merged, "alphabetagamma");

fn for_each<Func>(&mut self, func: Func) -> GeneratorResult where
Self: Sized,
Func: FnMut(Self::Output),

Run a generator to completion, or until it is stopped, and call a closure for each value produced by the generator.

The closure will be called for as long as the generator produces values, it is not possible to abort processing early. If early abort is needed, use Generator::run

Example

let mut sum = 0i32;
let data = [1,2,3];
let result = SliceGenerator::new(&data).for_each(|x| sum += x);
assert_eq!(sum, 6);
assert_eq!(result, GeneratorResult::Complete);

fn try_for_each<F, E>(&mut self, f: F) -> Result<(), E> where
Self: Sized,
F: FnMut(Self::Output) -> Result<(), E>,

A generator method that applies a fallible function to each item produced, stopping at the first error and returning that error.

This can also be thought of as the fallible form of for_each() or as the stateless version of try_fold().

Examples

use std::fs::rename;
use std::io::{stdout, Write};
use std::path::Path;
use pushgen::{SliceGenerator, GeneratorExt};

let data = ["no_tea.txt", "stale_bread.json", "torrential_rain.png"];

let res = SliceGenerator::new(&data).try_for_each(|x| writeln!(stdout(), "{}", x));
assert!(res.is_ok());

let mut gen = SliceGenerator::new(&data);
let res = gen.try_for_each(|x| rename(x, Path::new(x).with_extension("old")));
assert!(res.is_err());
// It short-circuited, so the remaining items are still in the iterator:
let mut output: Vec<&'static str> = Vec::new();
gen.for_each(|x| output.push(*x));
assert_eq!(output, ["stale_bread.json", "torrential_rain.png"]);

fn zip<Right>(self, right: Right) -> Zip<Self, Right> where
Self: Sized,
Right: Generator,

Zips the output of two generators into a single generator of pairs.

zip() returns a new generator that will use values from two generators, outputting a tuple where the first element comes from the first generator, and the second element comes from the second generator.

The zip generator will complete when either generator completes.

Example

use pushgen::{SliceGenerator, GeneratorExt};
let left = [1, 2, 3];
let right = [4, 5, 6];
let mut output: Vec<(i32, i32)> = Vec::new();
SliceGenerator::new(&left).zip(SliceGenerator::new(&right)).for_each(|(a, b)| output.push((*a, *b)));
assert_eq!(output, [(1,4), (2, 5), (3, 6)]);

fn dedup(self) -> Dedup<Self> where
Self: Sized,
Self::Output: PartialEq,

Create a de-duplicating generator, removing consecutive duplicate values.

Values will be made available when a non-duplicate is detected. If the up-stream generator generates the following sequence: [1, 2, 3, 3, 4] then the value 1 will be generated from the Dedup generator once the value 2 has been generated by the upstream generator and so on.

Upstream value	Dedup-generated value

1	None

2	1

3	2

3	Ignored

4	3

Complete	4

Complete	Complete

Example

let data = [1, 2, 3, 3, 3, 3, 4, 3];
let mut output: Vec<i32> = Vec::new();
SliceGenerator::new(&data).dedup().for_each(|x| output.push(*x));
assert_eq!(output, [1, 2, 3, 4, 3]);

fn iter(self) -> IteratorAdaptor<Self> where
Self: Sized,

Create an iterator from a generator.

This allows generators to be used in basic for-loops.

Example

use pushgen::{SliceGenerator, GeneratorExt};
let data = [1, 2, 3, 4, 5, 6];
let mut sum = 0;
for x in SliceGenerator::new(&data).iter() {
    sum += x;
}
assert_eq!(sum, data.iter().sum());

fn boxed(self) -> BoxedGenerator<Self::Output> where
Self: Sized + 'static,

Box a generator, making it possible to use as return value in for instance traits.

Performance

This causes at least one layer of redirection, which is very likely to impact performance. One should always prefer to use impl Generator<Output=X> instead.

Example

use pushgen::{BoxedGenerator, IntoGenerator, GeneratorExt};
fn make_generator() -> BoxedGenerator<i32> {
    vec![1, 2, 3, 4].into_gen().map(|x| x*2).boxed()
}
let mut output = Vec::new();
make_generator().for_each(|x| output.push(x));
assert_eq!(output, [2, 4, 6, 8]);

Loading content...